requirements.txt
# Machine learning
catboost==1.2.2
scikit-learn==1.3.1
numpy==1.26.0
pandas==2.1.1

# Web framework
fastapi>=0.68.0,<0.100.0
uvicorn[standard]>=0.15.0,<0.25.0

# Prometheus monitoring
prometheus-client>=0.14.1,<0.22.0
prometheus-fastapi-instrumentator>=6.0.0,<6.5.0

# Utilities
joblib==1.3.2
python-dotenv==1.1.1
requests>=2.31.0